Binds configuration to Configurable factory beans directly using Boot Binder.
Deprecates the use of Tuple.

Depracates ArgumentHints in favor of ShortcutConfigurable.

Creates Configurable interface to allow factories to create a
configuration object that can be bound to properties.

Creates StatefulConfigurable interface that extends Configurable and
allows singlton beans (like RedisRateLimiter) to store a map of
configuration keyed on route id. These beans use events like
FilterArgsEvent and PredicateArgsEvent to gather configuration as it is
read for each route.